# api/openapi.yaml
openapi: 3.0.0
info:
  description: This api is a test version to connect the backend and frontend
  title: Test api for project
  version: 1.0.0
servers:
- url: https://app.insertname.com/api
paths:
  /accounts/{accountID}:
    delete:
      description: Delete the account with this ID
      operationId: DeleteAccountByID
      parameters:
      - description: The ID of the account to return (UUID)
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: schema to delete the account with a specific id
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: delete_account_id
          type: string
        style: simple
      responses:
        "200":
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
    get:
      description: Returns an account by ID
      operationId: GetAccountByID
      parameters:
      - description: The ID of the account to return (UUID)
        example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: schema to get the account
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: schema_get_account_id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                ID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
                Name: Mark Wijnbergen
                Email: markwijnbergen@hey.com
              schema:
                $ref: '#/components/schemas/get_account_id_response'
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
  /accounts/{accountID}/finalize:
    post:
      description: finalize the account creation when user is added, this happens
        when email is senderAccountID
      operationId: FinalizeAccountCreation
      parameters:
      - description: The ID of the account
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: the schema to fetch the projects of a specific account
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: schema_projects_account_id
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/account_creation_finalize'
        content:
          application/json:
            example:
              emailAddress: markwijnbergen@hey.com
              firstName: mark
              lastName: Wijnbergen
              password: password123
            schema:
              properties:
                emailAddress:
                  description: email address of the participant
                  type: string
                firstName:
                  description: first name of the user
                  type: string
                lastName:
                  description: last name of the user
                  type: string
                password:
                  description: password of the user
                  type: string
              required:
              - emailAddress
              - firstName
              - lastName
              - password
              title: account_creation_finalize
              type: object
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: The specified account ID or email address is invalid
  /accounts/{accountID}/projects:
    get:
      description: returns all projects where this account is involved in
      operationId: GetProjectsOfAccount
      parameters:
      - description: The ID of the account
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: the schema to fetch the projects of a specific account
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: schema_projects_account_id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                projects:
                - projectID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
                  projectName: project 1
                - projectID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
                  projectName: project 2
              schema:
                $ref: '#/components/schemas/projects_account_id'
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
    post:
      description: Creates a new project
      operationId: CreatesNewProject
      parameters:
      - description: The ID of the account
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: the schema to fetch the projects of a specific account
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: schema_projects_account_id
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/create_project'
        content:
          application/json:
            example:
              projectTitle: Example Project
              participants:
              - firstmame: Mark
                lastname: Wijnbergen
                emailAddress: markwijnbergen@hey.com
              - firstname: John
                lastname: Doe
                emailAddress: johndoe@bla.com
              tests:
              - test1
              - test2
            schema:
              properties:
                projectTitle:
                  description: title of project
                  type: string
                participants:
                  description: participants of project
                  items:
                    $ref: '#/components/schemas/participant'
                  title: participants
                  type: array
                tests:
                  description: tests to perform as part of this project
                  items:
                    description: names of tests to perform
                    type: string
                  title: tests
                  type: array
              required:
              - participants
              - projectTitle
              - tests
              title: create_project
              type: object
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
  /authentication/login:
    post:
      description: logging in the account
      operationId: LogInWithAccount
      requestBody:
        $ref: '#/components/requestBodies/login_account'
        content:
          application/json:
            examples:
              password:
                value:
                  grantType: password
                  emailAddress: markwijnbergen@hey.com
                  password: password
              token:
                value:
                  grantType: refreshToken
                  emailAddress: markwijnbergen@hey.com
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50SUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJzZXNzaW9uSUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJleHBpcnlEYXRlIjoiMjAyMS0wMi0wNFQxNzowMTo0MC40NzRaIn0.JvCzrIHI60bm-Qekl3De0j4UEtQS-BykbDt9bhtmhLY
                  sessionID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2,
                  accountID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2,
            schema:
              properties:
                grantType:
                  description: type of grant type to log in with
                  enum:
                  - password
                  - refreshToken
                  type: string
                emailAddress:
                  description: account email address to log in with
                  type: string
                password:
                  description: password of the account to log in with
                  type: string
                refreshToken:
                  description: refreshToken of the account to log in with
                  type: string
                sessionID:
                  description: session of account wrl the refreshToken
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  type: string
                accountID:
                  description: id of the account
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  type: string
              required:
              - emailAddress
              - grantType
              title: login_account
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                accountID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
                sessionID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50SUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJzZXNzaW9uSUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJleHBpcnlEYXRlIjoiMjAyMS0wMi0wNFQxNzowMTo0MC40NzRaIn0.JvCzrIHI60bm-Qekl3De0j4UEtQS-BykbDt9bhtmhLY
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50SUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJzZXNzaW9uSUQiOiIwYThiYTlhZC1hOTNkLTQ5OTEtOGMyMy1jMTc4YjA5ZmU2YTIiLCJleHBpcnlEYXRlIjoiMjAyMS0wMi0wNFQxNzowMTo0MC40NzRaIn0.JvCzrIHI60bm-Qekl3De0j4UEtQS-BykbDt9bhtmhLY
                accountType: researcher
              schema:
                $ref: '#/components/schemas/Account_details'
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
  /authentication/refresh:
    post:
      description: Refresh the access token of the account
      operationId: RefreshAccessToken
      requestBody:
        $ref: '#/components/requestBodies/Refresh_details'
        content:
          application/json:
            example:
              password:
                value:
                  grantType: password
                  emailAddress: markwijnbergen@hey.com
                  password: password
            schema:
              description: Details necessary to refresh the access token
              properties:
                accountID:
                  description: The account ID.
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  type: string
                refreshToken:
                  description: the refreshToken used to update the acces token
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  type: string
              required:
              - accountID
              - refreshToken
              title: Refresh details
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                JWT: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50SUQiOiIxMjM0NTY3ODkwIiwibmFtZSI6Ik1hcmsgV2lqbmJlcmdlbiIsImFjY2Vzc3Rva2VuIjoxNTE2MjM5MDIyLCJyZWZyZXNodG9rZW4iOjc1Mzk3NTM1N30.V1JMLTr7R_yPdkbpZQ9kpUmmmq8cyEbpNRk5B78fpYw
              schema:
                $ref: '#/components/schemas/JWT_account_details'
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
  /authentication/revoke:
    post:
      description: logging out the account
      operationId: LogOutWithAccount
      requestBody:
        $ref: '#/components/requestBodies/logout_account'
        content:
          application/json:
            example:
              accessToken: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
            schema:
              description: schema to revoke access with the specific account
              properties:
                accessToken:
                  description: access token that needs to be revoked to log out
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  type: string
              required:
              - accessToken
              title: logout_account
              type: object
        required: true
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
  /projects/{projectID}/{accountID}/tests:
    get:
      description: Get all tests that need to be performed by account
      operationId: GetTestsToPerformByAccount
      parameters:
      - description: The ID of the project to return (UUID)
        explode: false
        in: path
        name: projectID
        required: true
        schema:
          description: the id of the project to get the tests from
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: project_id
          type: string
        style: simple
      - description: The ID of the account to return (UUID)
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: the id of the account to get the tests from
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: account_id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                testsToPerform:
                - testID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
                  testName: test 1
                - testID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a3
                  testName: test 2
              schema:
                $ref: '#/components/schemas/tests_project'
          description: OK
        "400":
          description: The specified project ID or account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
  /projects/{projectID}/genericTests/{testID}:
    get:
      description: Get a generic test with specific ID
      operationId: GetGenericTestOfProject
      parameters:
      - description: The ID of the project to return
        explode: false
        in: path
        name: projectID
        required: true
        schema:
          description: schema of the project id
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: project_id
          type: string
        style: simple
      - description: The ID of the test to return
        explode: false
        in: path
        name: testID
        required: true
        schema:
          description: schema of the test id
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: test_id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                title: Title of Questionnaire
                description: Hi thanks for joining
                displayAnswers: true
                finalRemark: Thanks for your participation
                questions:
                - question: whats your age
                  questionType: openQuestion
                  answers: {}
                - question: whats your sex
                  questionType: multipleChoice
                  answers:
                  - male
                  - female
                - question: rate your happines
                  questionType: slider
                  answers:
                  - very unhappy
                  - unhappy
                  - moderate
                  - happy
                  - very happy
              schema:
                $ref: '#/components/schemas/generic_test'
          description: OK
        "400":
          description: The specified project ID or test ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
    post:
      description: Submit the answer to the specific
      operationId: SubmitAnswerToTest
      parameters:
      - description: The ID of the project to return
        explode: false
        in: path
        name: projectID
        required: true
        schema:
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          type: string
        style: simple
      - description: The ID of the test to return (UUID)
        explode: false
        in: path
        name: testID
        required: true
        schema:
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/genericTestAnswers'
        content:
          application/json:
            example:
              accountID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
              answers:
              - question: 1
                answer: yo mama is so fat
                timeToAnswer: 5000
              - question: 2
                answer: They are taking the hobbits to Isengard
                timeToAnswer: 3999
            schema:
              description: The answers to a generic test
              properties:
                accountID:
                  description: the id of the account that performed this test
                  pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                  title: accountID
                  type: string
                answers:
                  description: the array of answers for this questionnaire
                  items:
                    $ref: '#/components/schemas/answer_to_question'
                  title: array of answers
                  type: array
              title: genericTestAnswers
              type: object
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: The specified project ID or test ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
  /test/{accountID}:
    get:
      description: Get all tests that are stored in the backend and are accessible
        for this account
      operationId: GetAllTests
      parameters:
      - description: The ID of the account to return (UUID)
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          description: the id of the account to get the tests from
          example: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          title: account_id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                tests:
                - testID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a1
                  testName: test 1
                - testID: 0a8ba9ad-a93d-4991-8c23-c178b09fe6a2
                  testName: test 2
              schema:
                description: all tests that researcher can address
                items:
                  $ref: '#/components/schemas/TestsForAccount'
                title: TestsForAccount
                type: array
          description: OK
        "400":
          description: The specified account ID is invalid
        "401":
          description: Authentication failed
        "404":
          description: Account not found
      security:
      - ApiKeyAuth: []
components:
  requestBodies:
    logout_account:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/logout_account'
      required: true
    create_project:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/create_project'
      required: true
    account_creation_finalize:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/account_creation_finalize'
      required: true
    genericTestAnswers:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/genericTestAnswers'
      required: true
    Refresh_details:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Refresh_details'
      required: true
    login_account:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/login_account'
      required: true
  schemas:
...
    participant:
      description: participant of project
      properties:
        firstame:
          description: firstname of the participant
          type: string
        lastname:
          description: lastname of the participant
          type: string
        emailAddress:
          description: email address of the participant
          type: string
      title: participant
      type: object
...
    JWT_account_details:
      description: the account details provided when the user logs in as a JWT token
      example:
        JWT: JWT
      properties:
        JWT:
          description: The jwt token with all information.
          type: string
      required:
      - JWT
      title: JWT account details
      type: object
...
    Account_details:
      description: the account details provided when the user logs in as a JWT token
      example:
        accountID: accountID
        accountType: researcher
        sessionID: sessionID
        accessToken: accessToken
        refreshToken: refreshToken
      properties:
        accountID:
          description: The account id.
          type: string
        sessionID:
          description: logged in SessionID
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          type: string
        accessToken:
          description: The access token.
          type: string
        refreshToken:
          description: The refresh token.
          type: string
        accountType:
          description: type of account
          enum:
          - researcher
          - participant
          type: string
      required:
      - accessToken
      - accountID
      - accountType
      - refreshToken
      - sessionID
      title: Account details
      type: object
    get_account_id_response:
      description: response to provide account details
      example:
        Email: Email
        ID: ID
        Name: Name
      properties:
        ID:
          description: The id of the account
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
          type: string
        Name:
          description: the name of the account
          type: string
        Email:
          description: the email address of the account
          type: string
      required:
      - Email
      - ID
      - Name
      title: get_account_id_response
      type: object
    projects_account_id:
      description: response with the projects of a specific account
      example:
        projects:
        - projectName: projectName
          projectID: projectID
        - projectName: projectName
          projectID: projectID
      properties:
        projects:
          description: an array with all projects
          items:
            $ref: '#/components/schemas/project'
          title: projects
          type: array
      title: projects_account_id
      type: object
  securitySchemes:
    ApiKeyAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http